requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-ratelimit==4.1.0
django-stubs==5.0.4
django-stubs-ext==5.0.4
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
tomli==2.0.1
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
urllib3==2.2.3
